pyproject.toml
[tool.poetry]
name = "kw6"
version = "0.0.0"
description = "Minimalistic library for reading files in the kw6 file format"
authors = ["NextML AB"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.2"
Pillow = "^9.0.0"
pydantic = "^2.0.0"

[tool.poetry.dev-dependencies]
tqdm = "^4.62.2"
pytest = "^6.2.5"
flake8 = "^3.9.2"
ipykernel = "^6.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
